Minor file manipulation
